Moved NavBar component to layout folder
