+++ client/src/App.jsx
@@ -1,6 +1,5 @@
-import { useState } from 'react'
 import './App.css'
-import NavBar from './NavBar'
+import NavBar from './assets/components/layout/NavBar'
 
 function App() {
