+++ client/src/assets/components/layout/NavBar.jsx
